// libs/common/src/schemas/mongoose/post/post.enum.ts
export enum PostEventsEnum {
  POST_SAVE_UPDATE_USER_COUNTS = 'post.post.save.updateUserCounts',
  POST_SAVE_UPDATE_PET_COUNTS = 'post.post.save.updatePetCounts',
  DELETE_DOC = 'post.deleteDoc',
  SUSPEND_DOC = 'post.suspendDoc',
  UN_SUSPEND_DOC = 'post.unSuspendDoc',
  USER_TAG_NOTIFICATION = 'post.userTag',
  PET_TAG_NOTIFICATION = 'post.petTag',
}